give julia a home
